great-expectations-wrapper/pyproject.toml
[tool.poetry]
name = "great_expectations_emr"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.6"
pyspark = "^2.4.5"
typer = "^0.1.1"
logzero = "^1.5.0"
boto3 = "^1.13.2"
pyyaml = "^5.3.1"
great_expectations = "0.12.4"
pypandoc = "^1.5"


[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
pytest = "^5.4.1"
pytest-mock = "^3.1.0"
freezegun = "^0.3.15"
mypy = "^0.770"
awscli = "^1.18.103"
black = "^19.10b0"
pylint = "^2.5.3"
rope = "^0.17.0"

[tool.poetry.scripts]
generate_dashboard = "great_expectations_emr.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
